requirements.txt
Flask
gunicorn
sqlalchemy
pymysql
python-dotenv
meteomatics
pandas